feat(app): add updated todo app and refactor code
