+++ app.py
@@ -1,4 +1,4 @@
-from app1 import app
+from core import app
 
 if __name__ == '__main__':
     app.run(debug=True)
