+++ core/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from app1 import db
+from core import db
 
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
