+++ core/templates/edit_note.html
