+++ core/templates/index.html
@@ -1,0 +1,106 @@
+{% extends "layout.html" %}
+{% block feed_body %}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <title>Notes</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <!-- Tailwind core -->
+  <script src="https://cdn.tailwindcss.com"></script>
+  <!-- Optional: Typography plugin via CDN for prose styles -->
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@tailwindcss/typography@0.2.0/dist/typography.min.css">
+  <script>
+    // Enable dark variant via data attribute or class
+    tailwind.config = {
+      darkMode: ['class', '[data-bs-theme="dark"]'],
+    }
+  </script>
+</head>
+<body class="bg-white text-zinc-900 dark:bg-zinc-950 dark:text-zinc-50">
+  <div class="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8 py-8">
+    <div class="grid grid-cols-1 lg:grid-cols-3 gap-6">
+      <!-- Left column -->
+      <div class="lg:col-span-1 space-y-6">
+        <div class="rounded-md border border-zinc-200 bg-white p-6 shadow-sm dark:border-zinc-800 dark:bg-zinc-900">
+          <h1 class="text-base font-semibold tracking-tight mb-4">View Notes</h1>
+          <form action="/view_notes" method="post" class="space-y-4">
+            <div>
+              <label for="year" class="block text-sm text-zinc-600 dark:text-zinc-300">Year</label>
+              <input type="number" name="year" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <div>
+              <label for="month" class="block text-sm text-zinc-600 dark:text-zinc-300">Month</label>
+              <input type="number" name="month" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <div>
+              <label for="day" class="block text-sm text-zinc-600 dark:text-zinc-300">Day</label>
+              <input type="number" name="day" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <button type="submit" class="inline-flex items-center rounded-md border border-zinc-300 bg-white px-3 py-1.5 text-sm text-zinc-800 shadow-sm hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950 dark:text-zinc-200 dark:hover:bg-zinc-900">
+              View
+            </button>
+          </form>
+        </div>
+
+        <div class="rounded-md border border-zinc-200 bg-white p-6 shadow-sm dark:border-zinc-800 dark:bg-zinc-900">
+          <h1 class="text-base font-semibold tracking-tight mb-4">Create Note</h1>
+          <form action="/create_note" method="post" class="space-y-4">
+            <div>
+              <label for="year" class="block text-sm text-zinc-600 dark:text-zinc-300">Year</label>
+              <input type="number" name="year" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <div>
+              <label for="month" class="block text-sm text-zinc-600 dark:text-zinc-300">Month</label>
+              <input type="number" name="month" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <div>
+              <label for="day" class="block text-sm text-zinc-600 dark:text-zinc-300">Day</label>
+              <input type="number" name="day" required class="mt-1 w-28 rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950" />
+            </div>
+            <div>
+              <label for="note" class="block text-sm text-zinc-600 dark:text-zinc-300">Note</label>
+              <textarea name="note" rows="4" required class="mt-1 block w-full rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950"></textarea>
+            </div>
+            <button type="submit" class="inline-flex items-center rounded-md bg-zinc-900 px-3 py-1.5 text-sm text-white shadow-sm hover:bg-zinc-800 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:bg-zinc-200 dark:text-zinc-900 dark:hover:bg-zinc-300">
+              Create
+            </button>
+          </form>
+        </div>
+      </div>
+
+      <!-- Right column -->
+      <div class="lg:col-span-2">
+        <div class="rounded-md border border-zinc-200 bg-white p-6 shadow-sm dark:border-zinc-800 dark:bg-zinc-900">
+          <h1 class="text-base font-semibold tracking-tight mb-4">Notes</h1>
+
+          {% if message %}
+          <p class="mt-2 text-sm text-zinc-600 dark:text-zinc-300">{{ message }}</p>
+          {% endif %}
+
+          {% if notes %}
+          <div class="space-y-4 max-h-[80vh] overflow-auto pr-2">
+            {% for note in notes %}
+            <div class="rounded-md border border-zinc-200 bg-zinc-50 p-4 dark:border-zinc-800 dark:bg-zinc-950/40">
+              <!-- Typography improvement for note body -->
+              <div class="prose prose-zinc max-w-none prose-p:my-2 prose-pre:bg-zinc-100 prose-code:px-1 prose-code:py-0.5 prose-code:bg-zinc-100 prose-code:rounded prose-headings:mt-3 prose-headings:mb-2 dark:prose-invert dark:prose-pre:bg-zinc-900 dark:prose-code:bg-zinc-900">
+                {{ note.note }}
+              </div>
+              <div class="mt-3 flex gap-2">
+                <a href="{{ url_for('edit_note', note_id=note.id) }}" class="inline-flex items-center rounded-md border border-zinc-300 bg-white px-3 py-1.5 text-sm text-zinc-800 shadow-sm hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-950 dark:text-zinc-200 dark:hover:bg-zinc-900">
+                  Edit
+                </a>
+                <a href="{{ url_for('delete_note', note_id=note.id) }}" class="inline-flex items-center rounded-md border border-red-300 bg-white px-3 py-1.5 text-sm text-red-700 shadow-sm hover:bg-red-50 focus:outline-none focus:ring-2 focus:ring-red-200 dark:border-red-700 dark:bg-zinc-950 dark:text-red-300 dark:hover:bg-red-900/30">
+                  Delete
+                </a>
+              </div>
+            </div>
+            {% endfor %}
+          </div>
+          {% endif %}
+        </div>
+      </div>
+    </div>
+  </div>
+</body>
+</html>
+{% endblock %}
