+++ core/templates/noteview.html
