+++ core/templates/todo_index.html
@@ -1,0 +1,99 @@
+{% extends 'layout.html' %}
+
+{% block feed_body %}
+<div class="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8 py-6">
+  <!-- Page header / toolbar -->
+  <div class="mb-4 flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
+    <div>
+      <h1 class="text-lg font-semibold tracking-tight text-zinc-900 dark:text-zinc-100">Task Master</h1>
+      <p class="text-xs text-zinc-500 dark:text-zinc-400">Overview of tasks and statuses</p>
+    </div>
+    <div class="flex w-full flex-col gap-2 sm:w-auto sm:flex-row sm:items-center">
+      <!-- Search (non-functional placeholder, safe to keep) -->
+      <div class="relative w-full sm:w-64">
+        <input type="search" placeholder="Search tasks…" class="w-full rounded-md border border-zinc-300 bg-white pl-9 pr-3 py-2 text-sm shadow-sm outline-none placeholder:text-zinc-400 focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-100" />
+        <svg class="pointer-events-none absolute left-2.5 top-2.5 h-4 w-4 text-zinc-400" viewBox="0 0 20 20" fill="currentColor"><path fill-rule="evenodd" d="M13.5 12a5.5 5.5 0 10-1.5 1.5l3.25 3.25a1 1 0 001.5-1.5L13.5 12zm-4 0a3.5 3.5 0 110-7 3.5 3.5 0 010 7z" clip-rule="evenodd"/></svg>
+      </div>
+      <div class="flex items-center gap-2">
+        <button type="button" class="inline-flex items-center gap-1 rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm text-zinc-800 shadow-sm hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-200 dark:hover:bg-zinc-800">
+          Filter
+        </button>
+        <button type="button" class="inline-flex items-center gap-1 rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm text-zinc-800 shadow-sm hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-200 dark:hover:bg-zinc-800">
+          Sort
+        </button>
+        <input type="date" class="rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-100" />
+      </div>
+    </div>
+  </div>
+
+  <!-- Table -->
+  <div class="overflow-hidden rounded-md border border-zinc-200 bg-white shadow-sm dark:border-zinc-800 dark:bg-zinc-900">
+    <div class="overflow-x-auto">
+      <table class="min-w-full table-auto text-sm">
+        <thead class="sticky top-0 z-10 bg-zinc-50/80 backdrop-blur dark:bg-zinc-900/70">
+          <tr class="text-left text-zinc-600 dark:text-zinc-300">
+            <th class="px-4 py-2 font-medium">Task no</th>
+            <th class="px-4 py-2 font-medium">Added</th>
+            <th class="px-4 py-2 font-medium">Creation Date</th>
+            <th class="px-4 py-2 font-medium">Completion Date</th>
+            <th class="px-4 py-2 font-medium">Status</th>
+            <th class="px-4 py-2 font-medium text-right">Actions</th>
+          </tr>
+        </thead>
+        <tbody class="divide-y divide-zinc-200 dark:divide-zinc-800">
+          {% for task in tasks %}
+          <tr class="hover:bg-zinc-50/70 dark:hover:bg-zinc-800/50">
+            <td class="px-4 py-2 text-zinc-800 dark:text-zinc-200">{{ task.id }}</td>
+            <td class="px-4 py-2 text-zinc-700 dark:text-zinc-300">
+              <span class="line-clamp-2 break-words">{{ task.content }}</span>
+            </td>
+            <td class="px-4 py-2 text-zinc-700 dark:text-zinc-300">{{ task.date_created.date() }}</td>
+            <td class="px-4 py-2 text-zinc-700 dark:text-zinc-300">
+              {% if task.completion_date %}
+                {{ task.completion_date.strftime("%Y-%m-%d") }}
+              {% else %}
+                <span class="text-zinc-500">Not completed</span>
+              {% endif %}
+            </td>
+            <td class="px-4 py-2">
+              {% if task.completion_time %}
+                <span class="inline-flex items-center gap-1 rounded-full bg-emerald-50 px-2.5 py-1 text-xs font-medium text-emerald-700 ring-1 ring-inset ring-emerald-200 dark:bg-emerald-900/20 dark:text-emerald-300 dark:ring-emerald-800">Completed</span>
+              {% else %}
+                <span class="inline-flex items-center gap-1 rounded-full bg-amber-50 px-2.5 py-1 text-xs font-medium text-amber-700 ring-1 ring-inset ring-amber-200 dark:bg-amber-900/20 dark:text-amber-300 dark:ring-amber-800">Pending</span>
+              {% endif %}
+            </td>
+            <td class="px-4 py-2">
+              <div class="flex items-center justify-end gap-2">
+                <a href="/delete/{{task.id}}" class="inline-flex items-center rounded-md border border-red-300 bg-white px-2.5 py-1.5 text-xs font-medium text-red-700 shadow-sm hover:bg-red-50 focus:outline-none focus:ring-2 focus:ring-red-200 dark:border-red-700 dark:bg-zinc-900 dark:text-red-300 dark:hover:bg-red-900/30">
+                  Delete
+                </a>
+                <a href="/update/{{task.id}}" class="inline-flex items-center rounded-md border border-zinc-300 bg-white px-2.5 py-1.5 text-xs font-medium text-zinc-800 shadow-sm hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-200 dark:hover:bg-zinc-800">
+                  Update
+                </a>
+                {% if not task.completion_time %}
+                <a href="/complete/{{task.id}}" class="inline-flex items-center rounded-md border border-emerald-300 bg-white px-2.5 py-1.5 text-xs font-medium text-emerald-700 shadow-sm hover:bg-emerald-50 focus:outline-none focus:ring-2 focus:ring-emerald-200 dark:border-emerald-700 dark:bg-zinc-900 dark:text-emerald-300 dark:hover:bg-emerald-900/30">
+                  Complete
+                </a>
+                {% endif %}
+              </div>
+            </td>
+          </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+    </div>
+  </div>
+
+  <!-- Add task input row -->
+  <form action="/index" method="POST" class="mt-6">
+    <div class="flex w-full items-center gap-2">
+      <input type="text" name="content" id="content" placeholder="Add Task" required
+             class="flex-1 rounded-md border border-zinc-300 bg-white px-3 py-2 text-sm text-zinc-900 placeholder:text-zinc-400 shadow-sm outline-none focus:border-zinc-400 focus:ring-2 focus:ring-zinc-300 dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-100" />
+      <button type="submit"
+              class="inline-flex items-center rounded-md bg-zinc-900 px-3 py-2 text-sm font-medium text-white shadow-sm hover:bg-zinc-800 focus:outline-none focus:ring-2 focus:ring-zinc-300 dark:bg-zinc-200 dark:text-zinc-900 dark:hover:bg-zinc-300">
+        Add Task
+      </button>
+    </div>
+  </form>
+</div>
+{% endblock %}
