+++ core/todo/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+from .todo import Todos
